requirements.txt
pymongo
python-mpd2
